requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
google-cloud-texttospeech
google-cloud-speech
mediapipe
opencv-python
numpy
pypdf
pydub
SpeechRecognition
python-multipart
websockets
gtts
